Require rspec core dependency explicitly (#10)
If Gemfile has rspec-rails without explicit rspec gem mention,
a Rails' Rake task doesn't load the rspec core, as the result
RSpec module gets defined, but ".configure" is absent. That
leads to the runtime error when saharspec gets loaded.

Load rspec-core explicitly to avoid the unexpected raise.
If no rspec-core in Gemfile it will raise an explicit error about that.